move ping/pong over to task manager
